Inject fault on update of articles
In order to test how well our tools detect faults, a forced
bug has been introduced. Now, updates to Articles appear to be
sucessful (since the updated article is returned), but GETting the
article will just return it as it was created. The testing tools
must have a way to do comparisons so this is detected.